Better divided Compare components
