+++ src/ducks/Studio/Compare/ProjectSelectDialog.module.scss
@@ -1,0 +1,12 @@
+.wrapper {
+  padding: 8px 16px;
+  width: 350px;
+}
+
+.search {
+  margin-bottom: 10px;
+}
+
+.projects {
+  height: 192px;
+}
