+++ src/ducks/Studio/Compare/Projects.module.scss
@@ -22,10 +22,6 @@
   padding: 8px;
 }
 
-.search {
-  margin-bottom: 10px;
-}
-
 .arrow {
   margin-left: 16px;
   fill: var(--waterloo);
