+++ src/ducks/Studio/Compare/utils.js
@@ -1,0 +1,3 @@
+export const projectSorter = ({ rank: a }, { rank: b }) => a - b
+
+export const hashComparable = ({ project, metric }) => project.slug + metric.key
